+++ snooze/plugins/core/notification/plugin.py
@@ -25,13 +25,14 @@ class Notification(Plugin):
         log.debug("Processing record {} against notifications".format(str(record.get('hash', ''))))
         for notification in self.notifications:
             if notification.enabled and notification.match(record):
-                log.debug("Notification {} matched record: {}".format(str(notification.name), str(record)))
+                log.debug("Notification {} matched record: {}".format(str(notification.name), str(record.get('hash', ''))))
                 if len(notification.action_plugins) > 0:
+                    actions_success = []
                     total = notification.freq.get('total', 1)
                     was_sent = None
                     if notification.freq.get('delay', 0) <= 0 and total != 0:
                         log.debug("{} Notification(s) `{}` will be sent now".format(total, notification.name))
-                        was_sent = notification.send(record)
+                        was_sent = notification.send(record, actions_success)
                     every = notification.freq.get('every', 0)
                     delay = max(notification.freq.get('delay', 0), 0) or every
                     retry = self.get_default(record, 'notification_retry', 3)
@@ -47,7 +48,7 @@ class Notification(Plugin):
                                 record['hash'] = hashlib.md5(record['raw']).hexdigest()
                             else:
                                 record['hash'] = hashlib.md5(repr(sorted(record.items())).encode('utf-8')).hexdigest()
-                        self.delay_send(notification, record['hash'], delay, every, total - 1, self.get_default(record, 'notification_retry', 3), was_sent)
+                        self.delay_send(notification, record['hash'], delay, every, total - 1, self.get_default(record, 'notification_retry', 3), was_sent, actions_success)
                 else:
                     log.error("Notification {} has no action. Cannot send".format(self.name))
         return record
@@ -63,12 +64,12 @@ class Notification(Plugin):
         '''Validate a notification object'''
         validate_condition(obj)
 
-    def delay_send(self, notification, record_hash, delay, every, total, retry, was_sent):
+    def delay_send(self, notification, record_hash, delay, every, total, retry, was_sent, actions_success):
         log.debug("Notification `{}` will be {}sent in {}s ({} retries left)".format(notification.name, '' if was_sent == True else 're', delay, retry))
         if was_sent == False:
             total += 1
-        self.thread.set_delayed(record_hash, notification.uid, {'notification': notification, 'time': time.time() + delay, 'every': every, 'total': total, 'retry': retry})
-        self.core.db.write('notification.delay', {'notification_uid': notification.uid, 'record_hash': record_hash, 'host': self.hostname, 'every': every, 'delay': delay, 'total': total, 'retry': retry}, 'notification_uid,record_hash')
+        self.thread.set_delayed(record_hash, notification.uid, {'notification': notification, 'time': time.time() + delay, 'every': every, 'total': total, 'retry': retry, 'actions_success': actions_success})
+        self.core.db.write('notification.delay', {'notification_uid': notification.uid, 'record_hash': record_hash, 'host': self.hostname, 'every': every, 'delay': delay, 'total': total, 'retry': retry, 'actions_success': actions_success}, 'notification_uid,record_hash')
 
     def post_init(self):
         super().post_init()
@@ -87,7 +88,8 @@ class Notification(Plugin):
                     every = delayed_notif['every']
                     total = delayed_notif['total']
                     retry = delayed_notif['retry']
-                    self.thread.set_delayed(record_hash, notif_uid, {'notification': notification, 'time': time.time() + delay, 'every': every, 'total': total, 'retry': retry})
+                    actions_success = delayed_notif['actions_success']
+                    self.thread.set_delayed(record_hash, notif_uid, {'notification': notification, 'time': time.time() + delay, 'every': every, 'total': total, 'retry': retry, 'actions_success': actions_success})
                 else:
                     log.debug("Delayed notification {} original notification in not in the database anymore. Removing it from queue".format(delayed_notif))
                     self.core.db.delete('notification.delay', ['=', 'uid', delayed_notif['uid']])
@@ -136,12 +138,12 @@ class NotificationObject():
         '''Whether a record match the Notification object'''
         return self.condition.match(record) and self.time_constraint.match(get_record_date(record))
 
-    def send_delayed(self, record_hash):
+    def send_delayed(self, record_hash, actions_success):
         delayed_records = self.core.db.search('record', ['=', 'hash', record_hash])
         if delayed_records['count'] > 0:
             for delayed_record in delayed_records['data']:
                 if delayed_record.get('state') not in ['ack', 'close']:
-                    was_sent = self.send(delayed_record)
+                    was_sent = self.send(delayed_record, actions_success)
                     self.core.db.write('record', delayed_record)
                     return (True, was_sent)
                 else:
@@ -151,31 +153,36 @@ class NotificationObject():
             log.debug("Record {} does not exist anymore, do not notify".format(record_hash))
             return (False, True)
 
-    def send(self, record):
-        for action_plugin in self.action_plugins:
-            if not 'notifications' in record:
-                record['notifications'] = []
+    def send(self, record, actions_success):
+        if not 'notifications' in record:
+            record['notifications'] = []
+        if self.name not in record['notifications']:
             record['notifications'].append(self.name)
+        was_sent = True
+        for action_plugin in self.action_plugins:
             action = action_plugin.get('action')
             action_content = action_plugin.get('content', {})
             action_name = action_content.get('action_name', '')
-            was_sent = False
-            try:
-                log.debug('Action {}'.format(action_plugin))
-                action.send(record, action_content)
-                was_sent = True
-            except Exception as e:
-                log.exception(e)
-                log.error("Notification {} action {}' could not be send".format(self.name, action_name))
-                was_sent = False
-            try:
-                if was_sent:
-                    self.core.stats.inc('notification_sent', {'name': self.name, 'action': action_name})
-                else:
-                    self.core.stats.inc('notification_error', {'name': self.name, 'action': action_name})
-            except Exception as e:
-                log.exception(e)
-            return was_sent
+            if action_name not in actions_success:
+                sent = False
+                try:
+                    log.debug('Action {}'.format(action_plugin))
+                    action.send(record, action_content)
+                    actions_success.append(action_name)
+                    sent = True
+                except Exception as e:
+                    log.exception(e)
+                    log.error("Notification {} action {}' could not be send".format(self.name, action_name))
+                    sent = False
+                    was_sent = False
+                try:
+                    if sent:
+                        self.core.stats.inc('notification_sent', {'name': self.name, 'action': action_name})
+                    else:
+                        self.core.stats.inc('notification_error', {'name': self.name, 'action': action_name})
+                except Exception as e:
+                    log.exception(e)
+        return was_sent
 
 
 class NotificationThread(threading.Thread):
@@ -199,12 +206,13 @@ class NotificationThread(threading.Thread):
                 for notif_uid in list(self.delayed[record_hash].keys()):
                     if time.time() >= self.delayed[record_hash][notif_uid]['time']:
                         notification = self.delayed[record_hash][notif_uid]['notification']
-                        keep, was_sent = notification.send_delayed(record_hash)
+                        actions_success = self.delayed[record_hash][notif_uid]['actions_success']
+                        keep, was_sent = notification.send_delayed(record_hash, actions_success)
                         every = max(self.delayed[record_hash][notif_uid]['every'], 0)
                         total = max(self.delayed[record_hash][notif_uid]['total'], -1)
                         retry = self.delayed[record_hash][notif_uid]['retry'] - 1
                         if (was_sent or retry > 0) and keep and every >= 0 and total != 0:
-                            self.notification.delay_send(notification, record_hash, every, every, total - 1, retry, was_sent)
+                            self.notification.delay_send(notification, record_hash, every, every, total - 1, retry, was_sent, actions_success)
                         else:
                             if keep:
                                 self.notification.core.db.delete('notification.delay', ['AND', ['=', 'record_hash', record_hash], ['=', 'notification_uid', notif_uid]])